make frontend versions dynamic
