+++ frontend/src/main.jsx
@@ -13,10 +13,17 @@ import { CollectionView } from './pages/CollectionView.jsx'
 import { PangenomeView } from './pages/PangenomeView.jsx'
 import { AttributeView } from './pages/AttributeView.jsx'
 import { DemoPage } from './pages/DemoPage.jsx'
-import { ComparisonView } from './pages/ComparisonView.jsx'
 import { SCIM } from './pages/SCIM.jsx'
 import { HomePage } from './pages/HomePage.jsx'
 import { HPRCGenomes } from './pages/HPRCGenomes.jsx'
+
+import { 
+  fetchServiceInfo,
+  fetchPangenomeLevels,
+  fetchSeqColList,
+  fetchCollectionLevels,
+  fecthComparison,
+} from './services/fetchData.jsx'
 
 import { AttributeValue, LinkedAttributeDigest } from './components/ValuesAndDigests.jsx'
 import { CollectionList, PangenomeList } from './components/ObjectLists.jsx'
@@ -34,73 +41,6 @@ import {
 
 
 import { API_BASE } from './utilities.jsx'
-
-const fetchPangenomeLevels = async(digest, level="2", collated=true) => {
-  const url = `${API_BASE}/pangenome/${digest}?level=1`
-  const url2 = `${API_BASE}/pangenome/${digest}?level=2`
-  const urlItemwise = `${API_BASE}/pangenome/${digest}?collated=false`  
-  let resps = [
-    fetch(url).then((response) => response.json()),
-    fetch(url2).then((response) => response.json()),
-    fetch(urlItemwise).then((response) => response.json())
-  ]
-  
-  return Promise.all(resps)
-}
-
-const fetchSeqColList = async() => {
-  const url = `${API_BASE}/list/collections?page_size=10&page=0`
-  const url2 = `${API_BASE}/list/pangenomes?page_size=5`
-  const url3 = `${API_BASE}/list/attributes/name_length_pairs?page_size=5`
-  let resps = [
-    fetch(url).then((response) => response.json()),
-    fetch(url2).then((response) => response.json()),
-    fetch(url3).then((response) => response.json())
-  ]
-  return Promise.all(resps)
-}
-
-const fetchSeqColDetails = async(digest, level="2", collated=true) => {
-  const url = `${API_BASE}/collection/${digest}?level=${level}&collated=${collated}`
-  return fetch(url).then((response) => response.json())
-}
-
-const fetchCollectionLevels = async (digest) => {
-  const urls = [
-    `${API_BASE}/collection/${digest}?level=1`,
-    `${API_BASE}/collection/${digest}?level=2`,
-    `${API_BASE}/collection/${digest}?collated=false`
-  ];
-
-  const responses = await Promise.all(urls.map(url =>
-    fetch(url).then(response => {
-      if (!response.ok) {
-        throw new Error(`Error fetching data from ${url}: ${response.statusText}`);
-      }
-      return response.json();
-    })
-  ));
-
-  return responses;
-};
-
-
-const fecthComparison = async(digest1, digest2) => {
-  const url = `${API_BASE}/comparison/${digest1}/${digest2}`
-  return fetch(url).then((response) => response.json())
-}
-
-const fetchAttribute = async(attribute, digest) => {
-  const url = `${API_BASE}/list/collections/${attribute}/${digest}`
-  const url2 = `${API_BASE}/attribute/collection/${attribute}/${digest}`
-  let resps = [
-    fetch(url).then((response) => response.json()),
-    fetch(url2).then((response) => response.json())
-  ]
-  return Promise.all(resps)
-}
-
-
 
 const Level1Collection = ({collection}) => {
   return (
@@ -124,8 +64,6 @@ const Level2Collection = ({collection}) => {
     </div>
   )
 }
-
-
 
 
 
@@ -158,7 +96,8 @@ const Nav = () => {
 }
 
 const App = () => { 
-  const refget_version = "0.5.0"
+  const loaderData = useLoaderData()
+  const refgetVersion = loaderData["version"]["refget_pkg_version"];
   return (<>
     <Nav/>
     <main className="container">
@@ -167,7 +106,10 @@ const App = () => {
     <div className="container">
       <footer className="flex-wrap py-3 my-4 align-top d-flex justify-content-between align-items-center border-top">
         <div className="d-flex flex-column"><div>
-          <span className="badge rounded-pill bg-primary text-primary bg-opacity-25 border border-primary me-1">refget {refget_version}</span>
+          <span className="badge rounded-pill bg-primary text-primary bg-opacity-25 border border-primary me-1">refget {refgetVersion}</span>
+          <span className="badge rounded-pill bg-primary text-primary bg-opacity-25 border border-primary me-1">seqcolapi {loaderData["version"]["seqcolapi_version"]}</span>
+          <span className="badge rounded-pill bg-primary text-primary bg-opacity-25 border border-primary me-1">python {loaderData["version"]["python_version"]}</span>
+          <span className="badge rounded-pill bg-primary text-primary bg-opacity-25 border border-primary me-1">seqcol spec {loaderData["version"]["seqcol_spec_version"]}</span>
           </div>
         <div className="d-flex flex-row mt-1 align-items-center">
           <div className="p-1 bg-success border border-success rounded-circle me-1"></div>
@@ -178,9 +120,6 @@ const App = () => {
     <ToastContainer />
   </>)
 }
-
-
-
 
 
 const CollectionTable = ({collections}) => {
@@ -210,8 +149,6 @@ const CollectionTable = ({collections}) => {
   )
 }
 
-
-
 function ErrorBoundary() {
   const error = useRouteError();
   console.error(error);
@@ -228,6 +165,7 @@ const router = createBrowserRouter([
   {
     path: "/",
     element: <App/>,
+    loader: fetchServiceInfo,
     children: [
       {
         path: "/",
